requirements.txt
iosense_connect
matplotlib
numpy
pandas
plotly
python-dotenv
Requests
streamlit
openpyxl
urllib3
seaborn
statsmodels
scikit-learn
pyparsing
uvicorn
celery
bentoml
mlflow
pydantic
fastapi
Requests
xgboost
tensorflow 
prophet 
pmdarima
imbalanced-learn
apache-airflow==2.6.3


